pyproject.toml
[tool.poetry]
name = "idanpow"
version = "0.1.0"
description = ""
authors = ["Te4nick <ivan.ivantz@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
numpy = "^1.26.4"
apscheduler = "^3.10.4"
gunicorn = "^22.0.0"
django-dotenv = "^1.4.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
